V1.1.2 完成薪資管理頁面、線材進貨頁面、線材生產頁面

+++ msystem/src/main/java/com/msystem/dto/GoodDto.java
@@ -38,7 +38,7 @@ public class GoodDto {
     private Date pDate;
 
     @JsonProperty("total_count")
-    private BigDecimal totalCount;
+    private int totalCount;
 
     @JsonProperty("unit_price")
     private BigDecimal unitPrice;
